requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
httpx==0.25.1
pydantic==2.7.4
langchain-core==0.3.68
langchain-openai==0.3.27
openai==1.86.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
gunicorn==21.2.0
requests==2.31.0
pillow==10.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# 视频生成相关依赖
manim==0.18.0
ffmpeg-python==0.2.0
celery==5.3.4
redis==5.0.1
# 腾讯云SDK (包含语音合成功能)
tencentcloud-sdk-python==3.0.1126
pydub==0.25.1
docker==7.0.0
prometheus-client==0.19.0
